// frontend/src/api/generated/model/secret.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Your API
 * OpenAPI spec version: 1.0.0
 */

export interface Secret {
  id?: number;
  name?: string;
  publicKey?: string;
  privateKey?: string;
  type?: string;
  lastRotatedAt?: string;
  active?: boolean;
  expiresAt?: string;
  createdAt?: string;
  updatedAt?: string;
}